Added Flexibility
Added more requirement functionality for types of degrees and field of degrees. Also updated the check job/edu functions to get a possible action space function. Will need work on the state to make fully function (specifically have education history be a list of their educations to account for having different types of degrees (masters of math + masters of econ)